+++ src/app/api/menu-items/route.js
@@ -1,38 +1,43 @@
-import { Category } from "@/app/models/Category";
 import { MenuItem } from "@/app/models/MenuItem";
 import mongoose from "mongoose";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req) {
     await mongoose.connect(process.env.MONGO_URL, {
         useNewUrlParser: true,
         useUnifiedTopology: true,
     });
-
     const data = await req.json();
 
-    // Ensure category exists and is a valid ObjectId
-    if (!data.category || !mongoose.Types.ObjectId.isValid(data.category)) {
-        return Response.json({ error: "Invalid or missing category ID" }, { status: 400 });
-    }
+    if (await isAdmin()) {
+        if (!data.category || !mongoose.Types.ObjectId.isValid(data.category)) {
+            return Response.json({ error: "Invalid or missing category ID" }, { status: 400 });
+        }
 
-    try {
-        const menuItemDoc = await MenuItem.create(data);
-        return Response.json(menuItemDoc);
-    } catch (error) {
-        return Response.json({ error: error.message }, { status: 500 });
+        try {
+            const menuItemDoc = await MenuItem.create(data);
+            return Response.json(menuItemDoc);
+        } catch (error) {
+            return Response.json({ error: error.message }, { status: 500 });
+        }
+    }
+    else{
+        return Response.json({});
     }
 }
 
 
 export async function PUT(req){
 
-        await mongoose.connect(process.env.MONGO_URL, {
-            useNewUrlParser: true,
-            useUnifiedTopology: true,
-        });  
+    await mongoose.connect(process.env.MONGO_URL, {
+        useNewUrlParser: true,
+        useUnifiedTopology: true,
+    });  
 
-    const {_id, ...data} = await req.json();
-    await MenuItem.findByIdAndUpdate(_id, data);
+    if (await isAdmin()) {
+        const {_id, ...data} = await req.json();
+        await MenuItem.findByIdAndUpdate(_id, data);
+    }
     return Response.json(true);
 
 }
@@ -43,9 +48,8 @@ export async function GET(){
             useNewUrlParser: true,
             useUnifiedTopology: true,
         });  
-        return Response.json(
-            await MenuItem.find()
-        )
+
+        return Response.json(await MenuItem.find())
 }
 
 export async function DELETE(req){
@@ -57,7 +61,8 @@ export async function DELETE(req){
 
     const url = new URL(req.url);
     const _id = url.searchParams.get('_id');
-    await MenuItem.deleteOne({_id});
-
+    if (await isAdmin()) {
+        await MenuItem.deleteOne({_id});
+    }
     return Response.json(true);
 }
